+++ musique/interpreter/interpreter.hh
@@ -25,10 +25,6 @@ struct Interpreter
 
 	std::function<std::optional<Error>(Interpreter&, Value)> default_action;
 
-	struct Incoming_Midi_Callbacks;
-	std::unique_ptr<Incoming_Midi_Callbacks> callbacks;
-	void register_callbacks();
-
 	Interpreter();
 	~Interpreter();
 	Interpreter(Interpreter const&) = delete;
@@ -57,7 +53,6 @@ struct Interpreter
 	void register_builtin_operators();
 };
 
-enum class Midi_Connection_Type { Output, Input };
-std::optional<Error> ensure_midi_connection_available(Interpreter&, Midi_Connection_Type, std::string_view operation_name);
+std::optional<Error> ensure_midi_connection_available(Interpreter&, std::string_view operation_name);
 
 #endif